+++ src/main/index.ts
@@ -23,7 +23,8 @@ async function createWindow(): Promise<BrowserWindow> {
       preload: join(__dirname, "../preload/index.cjs"),
       contextIsolation: true,
       nodeIntegration: false,
-      sandbox: true
+      sandbox: true,
+      webviewTag: true
     }
   });