+++ src/renderer/src/env.d.ts
@@ -8,4 +8,15 @@ declare global {
   }
 }
 
+declare module "react/jsx-runtime" {
+  namespace JSX {
+    interface IntrinsicElements {
+      webview: React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement> & {
+        src?: string;
+        partition?: string;
+      };
+    }
+  }
+}
+
 export {};